Send move input, broadcast and apply game state
Former-commit-id: 830f2e671b2bd881a5537add7ad048bec101678c

+++ server/packet.go
@@ -27,5 +27,5 @@ type Join struct {
 
 type Input struct {
     Id string
-    Move float64
+    Move int
 }
